@@ -51,7 +51,7 @@
     "lodash": "^4.17.21",
     "match-sorter": "^6.3.1",
     "moment": "^2.29.1",
-    "next": "13",
+    "next": "^13.4.19",
     "next-auth": "4.19.2",
     "next-remove-imports": "^1.0.6",
     "next-swagger-doc": "^0.3.4",
@@ -73,7 +73,8 @@
     "request": "^2.88.2",
     "swagger-ui-react": "^4.12.0",
     "use-debounce": "^9.0.4",
-    "yarn": "^1.22.17"
+    "yarn": "^1.22.17",
+    "database": "*"
   },
   "devDependencies": {
     "@types/node": "17.0.4",
