@@ -4,7 +4,7 @@ import { PrismaClient } from "@prisma/client";
 //   var prisma: PrismaClient | undefined;
 // }
 
-export const client = new PrismaClient();
+export const prisma = new PrismaClient();
 
 // if (process.env.NODE_ENV !== "production") global.prisma = prisma;
 
